pyproject.toml
[tool.poetry]
name = "books_app"
version = "0.1.0"
description = "A simple books app"
authors = ["Femi Ayo-Tubosun <femiayotubosun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
black = "^21.10b0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
coverage = "^6.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
